Break helper methods out into their own files for test command and move Test into 'testing' folder
