+++ packages/core/index.js
@@ -7,7 +7,7 @@ module.exports = {
   // TODO: update this to non-legacy the next breaking change
   contracts: require("@truffle/workflow-compile/legacy"),
   package: require("./lib/package"),
-  test: require("./lib/test"),
+  test: require("./lib/testing/test"),
   version: pkg.version,
   ganache: require("ganache-core/public-exports")
 };
